+++ app/src/main/java/com/infinity/omos/ui/onboarding/ErrorField.kt
@@ -1,4 +1,4 @@
-package com.infinity.omos.ui.onboarding.login
+package com.infinity.omos.ui.onboarding
 
 data class ErrorField(
     val state: Boolean,
